+++ src/AnonymousUtils.js
@@ -11,7 +11,7 @@
 
 import ParseUser from './ParseUser';
 import type { RequestOptions } from './RESTController';
-const uuidv4 = require('uuid/v4');
+const uuidv4 = require('./uuid');
 
 let registered = false;
