+++ src/InstallationController.js
@@ -10,12 +10,7 @@
  */
 
 import Storage from './Storage';
-let uuidv4 = null;
-if (process.env.PARSE_BUILD === 'weapp') {
-  uuidv4 = require('./uuid.weapp');
-} else {
-  uuidv4 = require('uuid/v4');
-}
+const uuidv4 = require('./uuid');
 
 let iidCache = null;
