+++ src/__tests__/RESTController-test.js
@@ -9,7 +9,7 @@
 
 jest.autoMockOff();
 jest.useFakeTimers();
-jest.mock('uuid/v4', () => {
+jest.mock('../uuid', () => {
   let value = 1000;
   return () => (value++).toString();
 });
